+++ coronavirus/apps.py
@@ -1,0 +1,8 @@
+from django.apps import AppConfig as DjangoApponfig
+
+
+class AppConfig(DjangoApponfig):
+    name = "coronavirus"
+    verbose_name = "Coronavirus"
+    include_in_administration_section = True
+    has_exportable_data = True