+++ meta_ae/choices.py
@@ -1,5 +1,21 @@
 from edc_constants.constants import UNKNOWN, OTHER
 
+AE_TYPE = (
+    ("sae", "Serious Adverse Event / Reaction"),
+    ("aesi", "Adverse Event of Special Interest"),
+    ("susar", "Serious Unexpected Adverse Reaction"),
+)
+
+AE_EXPECTED = (
+    ("expected", "Expected"),
+    ("unexpected", "Unexpected"),
+)
+
+AE_ACTION_REQUIRED = (
+    ("action", "Further action is required"),
+    ("no_action", "No further action is required"),
+)
+
 DEATH_LOCATIONS = (
     ("home", "At home"),
     ("hospital_clinic", "Hospital/clinic"),