+++ meta_ae/forms/ae_sponsor_review_form.py
@@ -1,0 +1,10 @@
+from django import forms
+
+from ..models import AeSponsorReview
+from .modelform_mixins import AeReviewModelFormMixin
+
+
+class AeSponsorReviewForm(AeReviewModelFormMixin, forms.ModelForm):
+    class Meta:
+        model = AeSponsorReview
+        fields = "__all__"