+++ meta_ae/models/__init__.py
@@ -1,5 +1,8 @@
 from .ae_followup import AeFollowup
 from .ae_initial import AeInitial
+
+# from .ae_local_review import AeLocalReview
+# from .ae_sponsor_review import AeSponsorReview
 from .ae_susar import AeSusar
 from .ae_tmg import AeTmg
 from .death_report import DeathReport