+++ meta_ae/models/ae_sponsor_review.py
@@ -1,0 +1,8 @@
+from edc_model import models as edc_models
+
+from ..model_mixins import AeReviewModelMixin
+
+
+class AeSponsorReview(AeReviewModelMixin, edc_models.BaseUuidModel):
+    class Meta:
+        verbose_name = "AE Sponsor Review"