+++ meta_auth/codenames/ae_local_reviewer.py
@@ -1,0 +1,7 @@
+ae_local_reviewer = [
+    "meta_subject.add_aelocalreview",
+    "meta_subject.change_aelocalreview",
+    "meta_subject.delete_aelocalreview",
+    "meta_subject.view_aelocalreview",
+    "meta_subject.view_historicalaelocalreview",
+]