+++ meta_auth/codenames/ae_sponsor_reviewer.py
@@ -1,0 +1,7 @@
+ae_sponsor_reviewer = [
+    "meta_subject.add_aesponsorreview",
+    "meta_subject.change_aesponsorreview",
+    "meta_subject.delete_aesponsorreview",
+    "meta_subject.view_aesponsorreview",
+    "meta_subject.view_historicalaesponsorreview",
+]