+++ meta_edc/urls.py
@@ -4,10 +4,10 @@ from django.contrib.auth.views import LogoutView
 from django.urls.conf import path, include
 from django.views.defaults import page_not_found, server_error  # noqa
 from django.views.generic.base import RedirectView
-
 from edc_action_item.admin_site import edc_action_item_admin
 from edc_adverse_event.admin_site import edc_adverse_event_admin
 from edc_appointment.admin_site import edc_appointment_admin
+from edc_crf.admin_site import edc_crf_admin
 from edc_dashboard.views import AdministrationView
 from edc_data_manager.admin_site import edc_data_manager_admin
 from edc_export.admin_site import edc_export_admin
@@ -54,6 +54,7 @@ urlpatterns = [
     path("admin/", admin.site.urls),
     path("admin/", edc_appointment_admin.urls),
     path("admin/", edc_adverse_event_admin.urls),
+    path("admin/", edc_crf_admin.urls),
     path("admin/", edc_randomization_admin.urls),
     path("admin/", meta_consent_admin.urls),
     path("admin/", meta_subject_admin.urls),
@@ -94,6 +95,7 @@ urlpatterns = [
     path("edc_adverse_event/", include("edc_adverse_event.urls")),
     path("edc_appointment/", include("edc_appointment.urls")),
     path("edc_action_item/", include("edc_action_item.urls")),
+    path("edc_crf/", include("edc_crf.urls")),
     path("edc_randomization/", include("edc_randomization.urls")),
     path("edc_dashboard/", include("edc_dashboard.urls")),
     path("edc_consent/", include("edc_consent.urls")),